Implemented functionality for employment rates

+++ packages/api/AlvTime.Business/Absence/VacationDaysDTO.cs
@@ -4,9 +4,9 @@ using AlvTime.Business.TimeEntries;
 namespace AlvTime.Business.Absence {
     public class VacationDaysDTO
     {
-        public int AvailableVacationDays { get; set; }    
-        public int PlannedVacationDaysThisYear { get; set; }    
-        public int UsedVacationDaysThisYear { get; set; }    
+        public decimal AvailableVacationDays { get; set; }    
+        public decimal PlannedVacationDaysThisYear { get; set; }    
+        public decimal UsedVacationDaysThisYear { get; set; }    
         public IEnumerable<TimeEntryResponseDto> PlannedTransactions { get; set; }
         public IEnumerable<TimeEntryResponseDto> UsedTransactions { get; set; }
     }
